Change: User registration process moved to AuthHandler
