+++ restapi/README.md
@@ -5,7 +5,7 @@
 | HTTPメソッド | パス             | 概要                                   |
 | :----------- | :--------------- | :------------------------------------- |
 | GET          | `/health`        | 疎通を確認する                         |
-| POST         | `/user` | ユーザーを登録する                     |
+| POST         | `/auth/register` | ユーザーを登録する                     |
 | GET          | `/user/hoge`     | ユーザー名が`hoge`のユーザーを取得する |
 | PUT          | `/user`   | ユーザーを更新する                     |
 | DELETE       | `/user/hoge`     | ユーザー名が`hoge`のユーザーを削除する |
