+++ restapi/handler/gen_moq_test.go
@@ -1,3 +1,3 @@
 package handler
 
-//go:generate go run github.com/matryer/moq -out moq_test.go . HealthService UserService
+//go:generate go run github.com/matryer/moq -out moq_test.go . HealthService UserService AuthService
