+++ restapi/handler/testdata/auth/register/email_exists_request.json.golden
