+++ restapi/handler/testdata/auth/register/email_exists_response.json.golden
