+++ restapi/handler/testdata/auth/register/empty_user_name_request.json.golden
