+++ restapi/handler/testdata/auth/register/empty_user_name_response.json.golden
