+++ restapi/handler/testdata/auth/register/ok_request.json.golden
