+++ restapi/handler/testdata/auth/register/ok_response.json.golden
