+++ restapi/handler/testdata/auth/register/user_name_exists_response.json.golden
