+++ restapi/handler/user_handler.go
@@ -11,7 +11,6 @@ import (
 )
 
 type UserService interface {
-	RegisterUser(ctx context.Context, data *model.UserRegistrationData) (*model.User, error)
 	GetUserByUserName(ctx context.Context, userName string) (*model.User, error)
 	UpdateUser(ctx context.Context, u *model.UserUpdateData) error
 	DeleteUserByUserName(ctx context.Context, userName string) error
@@ -19,35 +18,6 @@ type UserService interface {
 
 type UserHandler struct {
 	Service UserService
-}
-
-// POST /user
-// ユーザーを登録する
-func (uh *UserHandler) RegisterUser(c *gin.Context) {
-	var data *model.UserRegistrationData
-	if err := c.ShouldBindJSON(&data); err != nil {
-		c.Error(err)
-		c.AbortWithStatusJSON(http.StatusBadRequest, gin.H{"msg": BadRequestMessage})
-		return
-	}
-	u, err := uh.Service.RegisterUser(c.Request.Context(), data)
-	if err != nil {
-		c.Error(err)
-		// ユーザー名が既に登録されている場合
-		if errors.Is(err, service.ErrUserNameAlreadyExists) {
-			c.AbortWithStatusJSON(http.StatusBadRequest, gin.H{"msg": UserNameAlreadyEntryMessage})
-			return
-		}
-		// メールアドレスが既に登録されている場合
-		if errors.Is(err, service.ErrEmailAlreadyExists) {
-			c.AbortWithStatusJSON(http.StatusBadRequest, gin.H{"msg": EmailAlreadyEntryMessage})
-			return
-		}
-		c.AbortWithStatusJSON(http.StatusInternalServerError, gin.H{"msg": ServerErrorMessage})
-		return
-	}
-	// ユーザーIDを返す
-	c.JSON(http.StatusOK, gin.H{"id": u.Id})
 }
 
 // GET /user/:user_name
