+++ restapi/service/user.go
@@ -6,7 +6,6 @@ import (
 	"fmt"
 
 	"github.com/jmoiron/sqlx"
-	"github.com/kngnkg/tunetrail/restapi/auth"
 	"github.com/kngnkg/tunetrail/restapi/model"
 	"github.com/kngnkg/tunetrail/restapi/store"
 )
@@ -24,6 +23,7 @@ type UserRepository interface {
 	// WithTransactionはトランザクションを実行する
 	WithTransaction(ctx context.Context, db store.Beginner, f func(tx *sqlx.Tx) error) error
 	// RegisterUserはユーザーを登録する
+	// RegisterUserInfo
 	RegisterUser(ctx context.Context, db store.Queryer, u *model.User) error
 	// UserExistsByUserNameはユーザー名が既に存在するかどうかを返す
 	UserExistsByUserName(ctx context.Context, db store.Queryer, userName string) (bool, error)
@@ -35,56 +35,9 @@ type UserRepository interface {
 	DeleteUserByUserName(ctx context.Context, db store.Queryer, userName string) error
 }
 
-type Auth interface {
-	SignUp(ctx context.Context, email, password string) (string, error)
-}
-
 type UserService struct {
 	DB   store.Beginner
 	Repo UserRepository
-	Auth Auth
-}
-
-// RegisterUserはユーザーを登録する
-func (us *UserService) RegisterUser(ctx context.Context, data *model.UserRegistrationData) (*model.User, error) {
-	id, err := us.Auth.SignUp(ctx, data.Email, data.Password)
-	if err != nil {
-		if errors.Is(err, auth.ErrEmailAlreadyExists) {
-			return nil, fmt.Errorf("%w: %w", ErrEmailAlreadyExists, err)
-		}
-		return nil, err
-	}
-
-	registeredUser := &model.User{}
-	// トランザクション開始
-	err = us.Repo.WithTransaction(ctx, us.DB, func(tx *sqlx.Tx) error {
-		// ユーザー名が既に存在するかどうかを確認
-		exists, err := us.Repo.UserExistsByUserName(ctx, tx, data.UserName)
-		if err != nil {
-			return err
-		}
-		if exists {
-			return ErrUserNameAlreadyExists
-		}
-
-		u := &model.User{
-			Id:       id,
-			UserName: data.UserName,
-			Name:     data.Name,
-			IconUrl:  "",
-			Bio:      "",
-		}
-		if err = us.Repo.RegisterUser(ctx, tx, u); err != nil {
-			return err
-		}
-		registeredUser = u
-		return nil
-	})
-	if err != nil {
-		return nil, err
-	}
-
-	return registeredUser, nil
 }
 
 // GetUserByUserNameはユーザー名からユーザーを取得する
