+++ restapi/service/user_test.go
@@ -8,7 +8,6 @@ import (
 	"testing"
 
 	"github.com/jmoiron/sqlx"
-	"github.com/kngnkg/tunetrail/restapi/auth"
 	"github.com/kngnkg/tunetrail/restapi/clock"
 	"github.com/kngnkg/tunetrail/restapi/model"
 	"github.com/kngnkg/tunetrail/restapi/store"
@@ -43,7 +42,6 @@ func TestUserServiceTestSuite(t *testing.T) {
 	t.Parallel()
 	moqDB := &BeginnerMock{}
 	moqRepo := &UserRepositoryMock{}
-	moqAuth := &AuthMock{}
 	fc := &clock.FixedClocker{}
 
 	// テスト用のダミーユーザーを作成
@@ -88,23 +86,8 @@ func TestUserServiceTestSuite(t *testing.T) {
 		return nil
 	}
 
-	moqRepo.RegisterUserFunc = func(ctx context.Context, db store.Queryer, u *model.User) error {
-		// ダミーの値を設定
-		u.Id = "1"
-		u.CreatedAt = fc.Now()
-		u.UpdatedAt = fc.Now()
-		return nil
-	}
-
 	moqRepo.UserExistsByUserNameFunc = func(ctx context.Context, db store.Queryer, userName string) (bool, error) {
 		if userName == DUMMY_1_USERNAME || userName == DUMMY_2_USERNAME {
-			return true, nil
-		}
-		return false, nil
-	}
-
-	moqRepo.UserExistsByEmailFunc = func(ctx context.Context, db store.Queryer, email string) (bool, error) {
-		if email == DUMMY_1_EMAIL || email == DUMMY_2_EMAIL {
 			return true, nil
 		}
 		return false, nil
@@ -139,24 +122,10 @@ func TestUserServiceTestSuite(t *testing.T) {
 		return store.ErrUserNotFound
 	}
 
-	moqAuth.SignUpFunc = func(ctx context.Context, email, password string) (string, error) {
-		if password != VALID_PASSWORD {
-			return "", auth.ErrInvalidPassword
-		}
-		if password == DUMMY_1_PASSWORD || password == DUMMY_2_PASSWORD {
-			return "", auth.ErrInvalidPassword
-		}
-		if email == DUMMY_1_EMAIL || email == DUMMY_2_EMAIL {
-			return "", auth.ErrEmailAlreadyExists
-		}
-		return VALID_USER_ID, nil
-	}
-
 	suite.Run(t, &UserServiceTestSuite{
 		us: &UserService{
 			DB:   moqDB,
 			Repo: moqRepo,
-			Auth: moqAuth,
 		},
 		fc:         fc,
 		dummyUsers: dummyUsers,
@@ -164,86 +133,6 @@ func TestUserServiceTestSuite(t *testing.T) {
 }
 
 func (s *UserServiceTestSuite) SetupTest() {}
-
-func (s *UserServiceTestSuite) TestRegisterUser() {
-	type args struct {
-		ctx  context.Context
-		data *model.UserRegistrationData
-	}
-
-	tests := []struct {
-		name     string
-		args     args
-		wantUser *model.User
-		wantErr  error
-	}{
-		{
-			"ok",
-			args{
-				ctx: context.Background(),
-				data: &model.UserRegistrationData{
-					UserName: "test",
-					Name:     "test",
-					Password: VALID_PASSWORD,
-					Email:    "test@example.com",
-				},
-			},
-			&model.User{
-				Id:        VALID_USER_ID,
-				UserName:  "test",
-				Name:      "test",
-				IconUrl:   "",
-				Bio:       "",
-				CreatedAt: s.fc.Now(),
-				UpdatedAt: s.fc.Now(),
-			},
-			nil,
-		},
-		{
-			// ユーザー名が重複している場合
-			"errUserNameAlreadyExists",
-			args{
-				ctx: context.Background(),
-				data: &model.UserRegistrationData{
-					UserName: DUMMY_1_USERNAME, // ダミーユーザーのユーザー名と重複させる
-					Name:     "test",
-					Password: VALID_PASSWORD,
-					Email:    "test@example.com",
-				},
-			},
-			nil,
-			ErrUserNameAlreadyExists,
-		},
-		{
-			// メールアドレスが重複している場合
-			"errEmailAlreadyExists",
-			args{
-				ctx: context.Background(),
-				data: &model.UserRegistrationData{
-					UserName: "test",
-					Name:     "test",
-					Password: VALID_PASSWORD,
-					Email:    DUMMY_1_EMAIL, // ダミーユーザーのメールアドレスと重複させる
-				},
-			},
-			nil,
-			ErrEmailAlreadyExists,
-		},
-		// パスワードが不正な場合
-	}
-
-	for _, tt := range tests {
-		s.Run(tt.name, func() {
-			got, err := s.us.RegisterUser(tt.args.ctx, tt.args.data)
-			if !errors.Is(err, tt.wantErr) {
-				s.T().Errorf("UserService.RegisterUser() error = %v, wantErr %v", err, tt.wantErr)
-			}
-			if tt.wantErr == nil {
-				assert.Equal(s.T(), tt.wantUser, got)
-			}
-		})
-	}
-}
 
 func (s *UserServiceTestSuite) TestGetUserByUserName() {
 	type args struct {
